Use the indented table-of-content js code from the doc view for the righthand-nav
