+++ ietf/static/js/document_html.js
@@ -8,14 +8,9 @@ import {
 
 import debounce from "lodash/debounce";
 import Cookies from "js-cookie";
+import { populate_nav } from "./nav.js";
 
 const cookies = Cookies.withAttributes({ sameSite: "strict" });
-
-function make_nav() {
-    const nav = document.createElement("nav");
-    nav.classList.add("nav-pills", "ps-3", "flex-column");
-    return nav;
-}
 
 // Chrome apparently wants this debounced to something >10ms,
 // otherwise the main view doesn't scroll?
@@ -61,73 +56,7 @@ document.addEventListener("DOMContentLoaded", function (event) {
 
     // Set up a nav pane
     const toc_pane = document.querySelector("#toc-nav");
-
-    // Extract section headings from document
-    const headings = document.querySelectorAll(
-        // "#content h1, "
-        "#content h2, #content h3, " +
-        "#content h4, #content h5, #content h6"
-    );
-
-    let nav_stack = [toc_pane];
-    let cur_level = 0;
-    headings.forEach(el => {
-        let level = el.tagName.charAt(el.tagName.length - 1) -
-            2;
-        if (level < cur_level) {
-            while (level < cur_level) {
-                let nav = nav_stack.pop();
-                cur_level--;
-                nav_stack[level].appendChild(nav);
-            }
-        } else {
-            while (level > cur_level) {
-                nav_stack.push(make_nav());
-                cur_level++;
-            }
-        }
-
-        const link = document.createElement("a");
-        link.classList.add("nav-link", "py-0", "ps-1", "d-flex");
-
-        for (let i = 0; i < el.childNodes.length; i++) {
-            if (el.childNodes[i].attributes &&
-                el.childNodes[i].attributes.href.value) {
-                link.href = el.childNodes[i].attributes.href.value;
-            }
-        }
-        if (!link.href) {
-            link.classList.add("disabled");
-            link.setAttribute("href", "#");
-        }
-
-        const words = el.innerText.split(/\s+/);
-        let nr = "";
-        if (words[0].includes(".")) {
-            nr = words.shift();
-        } else if (words.length > 1 && words[1].includes(".")) {
-            nr = words.shift() + " " + words.shift();
-            nr = nr.replace(/\s*Appendix\s*/, "");
-        }
-
-        if (nr) {
-            const number = document.createElement("div");
-            number.classList.add("pe-1");
-            number.textContent = nr;
-            link.appendChild(number);
-        }
-
-        const text = document.createElement("div");
-        text.classList.add("text-break");
-        text.textContent = words.join(" ");
-        link.appendChild(text);
-
-        nav_stack[level].appendChild(link);
-    });
-
-    for (var i = nav_stack.length - 1; i > 0; i--) {
-        nav_stack[i - 1].appendChild(nav_stack[i]);
-    }
+    populate_nav(toc_pane, "#content h2, #content h3, #content h4, #content h5, #content h6", 2, ["py-0"]);
 
     // activate pref buttons selected by pref cookies
     document.querySelectorAll(".btn-check")
